# docs/api.yaml
openapi: 3.0.3
info:
  title: Attendify API
  version: 1.0.0
  description: |
    Attendify is a secure, distributed, cloud-native SaaS backend system.

    The platform provides:

      - Company identity management
      - Secure multi-tenant authentication
      - Edge-based request routing through Cloudflare Worker
      - JWT-based stateless authentication
      - Cryptographic request verification using HMAC-SHA256
      - Nonce-based replay protection
      - Secure attendance evidence submission

    Security model:

      - Clients communicate with the Cloudflare Worker gateway.
      - The Cloudflare Worker forwards requests to the Attendify backend.
      - The backend validates trusted gateway traffic using an internal edge secret.
      - Protected company routes require JWT Bearer authentication.
      - Critical attendance requests require nonce freshness and HMAC signature verification.

    Request flow:

      Client -> Cloudflare Worker -> Attendify Backend -> MongoDB Atlas

    Important production rule:

      Clients should use the Cloudflare Worker URL as the public API entry point.
      The Railway backend URL should not be used directly by clients.

  license:
    name: Private / Internal Use
servers:
  - url: http://localhost:3000
    description: Local Development Server

  - url: https://attendify-worker.mahmouddamour-3aa.workers.dev
    description: Production Edge Gateway
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      operationId: registerCompany
      summary: Register a new company
      description: |
        Creates a new company tenant account.

        Security behavior:

          - Email is normalized before storage.
          - Password is hashed using bcrypt.
          - Duplicate company email registration is rejected.
          - Plaintext passwords are never stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              validRegister:
                summary: Valid company registration
                value:
                  name: acme
                  email: admin@acme.com
                  password: strongpassword
      responses:
        "200":
          description: Company registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                success: true
                message: Company registered successfully

        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Invalid input data

        "409":
          description: Company already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Company already exists

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Internal server error
  /auth/login:
    post:
      tags:
        - Auth
      operationId: loginCompany
      summary: Login and receive JWT token
      description: |
        Authenticates a company using email and password.

        On successful authentication:

          - A signed JWT token is returned.
          - The client must send the token in the Authorization header.
          - The token follows the Bearer authentication scheme.

        Header usage:

          Authorization: Bearer <JWT>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              validLogin:
                summary: Valid login request
                value:
                  email: admin@acme.com
                  password: strongpassword
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                success: true
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

        "400":
          description: Invalid credentials format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Invalid credentials

        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Invalid credentials

        "403":
          description: Account temporarily locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Account temporarily locked

        "404":
          description: Company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Company not found

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication.

        Usage:

          Authorization: Bearer <token>
  schemas:
...
    SuccessMessage:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Internal server error
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          example: acme
        email:
          type: string
          format: email
          example: admin@acme.com
        password:
          type: string
          format: password
          minLength: 6
          example: strongpassword
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@acme.com
        password:
          type: string
          format: password
          minLength: 6
          example: strongpassword
    AuthResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: Signed JWT token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...